fix: verifier end state

+++ packages/legacy/core/App/screens/ProofDetails.tsx
@@ -9,7 +9,6 @@ import {
   markProofAsViewed,
 } from '@hyperledger/aries-bifold-verifier'
 import { useFocusEffect } from '@react-navigation/native'
-import { StackNavigationProp } from '@react-navigation/stack'
 import React, { useCallback, useEffect, useMemo, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import { BackHandler, ScrollView, StyleSheet, Text, View } from 'react-native'
@@ -25,28 +24,32 @@ import { testIdWithKey } from '../utils/testable'
 
 type ProofDetailsProps = StackScreenProps<ProofRequestsStackParams, Screens.ProofDetails>
 
-interface VerifiedProofProps {
+type VerifiedProofProps = {
   record: ProofExchangeRecord
   isHistory?: boolean
   senderReview?: boolean
-  navigation: StackNavigationProp<ProofRequestsStackParams, Screens.ProofDetails>
+  connectionLabel: string
+  onBackPressed: () => void
+  onGenerateNewPressed: () => void
 }
 
-interface UnverifiedProofProps {
+type UnverifiedProofProps = {
   record: ProofExchangeRecord
-  navigation: StackNavigationProp<ProofRequestsStackParams, Screens.ProofDetails>
+  onBackPressed: () => void
+  onGenerateNewPressed: () => void
 }
 
 const VerifiedProof: React.FC<VerifiedProofProps> = ({
   record,
-  navigation,
   isHistory,
   senderReview,
+  connectionLabel,
+  onBackPressed,
+  onGenerateNewPressed,
 }: VerifiedProofProps) => {
   const { t } = useTranslation()
-  const { ColorPallet, TextTheme, Assets } = useTheme()
-  const [store] = useStore()
-
+  const { ColorPallet, TextTheme } = useTheme()
+  const [sharedProofDataItems, setSharedProofDataItems] = useState<GroupedSharedProofDataItem[]>([])
   const styles = StyleSheet.create({
     container: {
       flexGrow: 1,
@@ -99,38 +102,9 @@ const VerifiedProof: React.FC<VerifiedProofProps> = ({
     },
   })
 
-  const connection = useConnectionById(record.connectionId || '')
-  const connectionLabel = useMemo(
-    () =>
-      connection
-        ? getConnectionName(connection, store.preferences.alternateContactNames)
-        : t('Verifier.ConnectionLessLabel'),
-    [connection, store.preferences.alternateContactNames, t]
-  )
-
-  const [sharedProofDataItems, setSharedProofDataItems] = useState<GroupedSharedProofDataItem[]>([])
-
   const onSharedProofDataLoad = useCallback((data: GroupedSharedProofDataItem[]) => {
     setSharedProofDataItems(data)
   }, [])
-
-  const onGenerateNew = useCallback(() => {
-    const metadata = record.metadata.get(ProofMetadata.customMetadata) as ProofCustomMetadata
-    if (metadata?.proof_request_template_id) {
-      navigation.navigate(Screens.ProofRequesting, { templateId: metadata.proof_request_template_id })
-    } else {
-      navigation.navigate(Screens.ProofRequests, {})
-    }
-  }, [record, navigation])
-
-  const onBack = useCallback(() => {
-    navigation.navigate(Screens.ProofRequests, {})
-  }, [navigation])
-
-  useEffect(() => {
-    if (!connection || !isHistory) return
-    navigation.setOptions({ title: connectionLabel })
-  }, [connection, isHistory, navigation, connectionLabel])
 
   if (isHistory) {
     return (
@@ -165,9 +139,7 @@ const VerifiedProof: React.FC<VerifiedProofProps> = ({
     <ScrollView contentContainerStyle={{ flexGrow: 1 }} testID={testIdWithKey('ProofDetailsView')}>
       <View style={styles.container}>
         <View style={styles.header}>
-          <View style={styles.iconContainer}>
-            <Assets.svg.informationReceived />
-          </View>
+          <View style={styles.iconContainer}>{/* <Assets.svg.informationReceived /> */}</View>
           <Text>
             <Text style={styles.headerTitle}>{t('Verifier.InformationReceived') + ' '}</Text>
             <Text style={styles.headerDetails}>{t('Verifier.InformationReceivedDetails')}</Text>
@@ -183,7 +155,7 @@ const VerifiedProof: React.FC<VerifiedProofProps> = ({
               accessibilityLabel={t('Verifier.GenerateNewQR')}
               testID={testIdWithKey('GenerateNewQR')}
               buttonType={ButtonType.Primary}
-              onPress={onGenerateNew}
+              onPress={onGenerateNewPressed}
             />
           </View>
           <Button
@@ -191,7 +163,7 @@ const VerifiedProof: React.FC<VerifiedProofProps> = ({
             accessibilityLabel={t('Verifier.BackToList')}
             testID={testIdWithKey('BackToList')}
             buttonType={ButtonType.Secondary}
-            onPress={onBack}
+            onPress={onBackPressed}
           />
         </View>
       </View>
@@ -199,10 +171,9 @@ const VerifiedProof: React.FC<VerifiedProofProps> = ({
   )
 }
 
-const UnverifiedProof: React.FC<UnverifiedProofProps> = ({ record, navigation }) => {
+const UnverifiedProof: React.FC<UnverifiedProofProps> = ({ record, onBackPressed, onGenerateNewPressed }) => {
   const { t } = useTranslation()
   const { TextTheme, Assets } = useTheme()
-
   const styles = StyleSheet.create({
     header: {
       paddingHorizontal: 20,
@@ -227,19 +198,6 @@ const UnverifiedProof: React.FC<UnverifiedProofProps> = ({ record, navigation })
     },
   })
 
-  const onGenerateNew = useCallback(() => {
-    const metadata = record.metadata.get(ProofMetadata.customMetadata) as ProofCustomMetadata
-    if (metadata?.proof_request_template_id) {
-      navigation.navigate(Screens.ProofRequesting, { templateId: metadata.proof_request_template_id })
-    } else {
-      navigation.navigate(Screens.ProofRequests, {})
-    }
-  }, [record, navigation])
-
-  const onBackToList = useCallback(() => {
-    navigation.navigate(Screens.ProofRequests, {})
-  }, [navigation])
-
   return (
     <ScrollView contentContainerStyle={{ flexGrow: 1 }} testID={testIdWithKey('UnverifiedProofView')}>
       <View style={styles.header}>
@@ -260,7 +218,7 @@ const UnverifiedProof: React.FC<UnverifiedProofProps> = ({ record, navigation })
             accessibilityLabel={t('Verifier.GenerateNewQR')}
             testID={testIdWithKey('GenerateNewQR')}
             buttonType={ButtonType.Primary}
-            onPress={onGenerateNew}
+            onPress={onGenerateNewPressed}
           />
         </View>
         <View style={styles.buttonContainer}>
@@ -269,7 +227,7 @@ const UnverifiedProof: React.FC<UnverifiedProofProps> = ({ record, navigation })
             accessibilityLabel={t('Verifier.BackToList')}
             testID={testIdWithKey('BackToList')}
             buttonType={ButtonType.Secondary}
-            onPress={onBackToList}
+            onPress={onBackPressed}
           />
         </View>
       </View>
@@ -284,19 +242,60 @@ const ProofDetails: React.FC<ProofDetailsProps> = ({ route, navigation }) => {
 
   const { recordId, isHistory, senderReview } = route.params
   const record = useProofById(recordId)
+  const connection = useConnectionById(record?.connectionId ?? '')
+  const { t } = useTranslation()
   const { agent } = useAgent()
   const [store] = useStore()
 
-  useEffect(() => {
-    return () => {
-      if (!store.preferences.useDataRetention) {
-        agent?.proofs.deleteById(recordId)
-      }
-      if ((record?.metadata.get(ProofMetadata.customMetadata) as ProofCustomMetadata).delete_conn_after_seen) {
-        agent?.connections.deleteById(record?.connectionId ?? '')
-      }
+  const connectionLabel = useMemo(
+    () =>
+      connection
+        ? getConnectionName(connection, store.preferences.alternateContactNames)
+        : t('Verifier.ConnectionLessLabel'),
+    [connection, store.preferences.alternateContactNames, t]
+  )
+
+  const cleanup = useCallback((): Promise<void> | undefined => {
+    if (!store.preferences.useDataRetention) {
+      return agent?.proofs.deleteById(recordId)
+    }
+
+    if (record && (record.metadata.get(ProofMetadata.customMetadata) as ProofCustomMetadata).delete_conn_after_seen) {
+      return agent?.connections.deleteById(record?.connectionId ?? '')
     }
   }, [store.preferences.useDataRetention, agent, recordId, record])
+
+  const onBackPressed = useCallback(() => {
+    cleanup()?.catch((err) => console.error(err))
+
+    if (route.params.isHistory) {
+      navigation.goBack()
+      return null
+    }
+
+    if (connection) {
+      navigation.setOptions({ title: connectionLabel })
+    }
+
+    navigation.navigate(Screens.ProofRequests, {})
+
+    return null
+  }, [navigation])
+
+  const onGenerateNewPressed = useCallback(() => {
+    if (!record) {
+      return
+    }
+
+    cleanup()?.catch((err) => console.error(err))
+
+    const metadata = record.metadata.get(ProofMetadata.customMetadata) as ProofCustomMetadata
+    if (metadata?.proof_request_template_id) {
+      navigation.navigate(Screens.ProofRequesting, { templateId: metadata.proof_request_template_id })
+    } else {
+      navigation.navigate(Screens.ProofRequests, {})
+    }
+  }, [record, navigation])
 
   useEffect(() => {
     if (agent && record && !record.metadata?.data?.customMetadata?.details_seen) {
@@ -306,18 +305,9 @@ const ProofDetails: React.FC<ProofDetailsProps> = ({ route, navigation }) => {
 
   useFocusEffect(
     useCallback(() => {
-      const onBackPress = () => {
-        if (route.params.isHistory) {
-          navigation.goBack()
-        } else {
-          navigation.navigate(Screens.ProofRequests, {})
-        }
-        return true
-      }
+      BackHandler.addEventListener('hardwareBackPress', onBackPressed)
 
-      BackHandler.addEventListener('hardwareBackPress', onBackPress)
-
-      return () => BackHandler.removeEventListener('hardwareBackPress', onBackPress)
+      return () => BackHandler.removeEventListener('hardwareBackPress', onBackPressed)
     }, [route.params, navigation])
   )
 
@@ -326,9 +316,18 @@ const ProofDetails: React.FC<ProofDetailsProps> = ({ route, navigation }) => {
   return (
     <SafeAreaView style={{ flexGrow: 1 }} edges={['left', 'right']}>
       {(record.isVerified || senderReview) && (
-        <VerifiedProof record={record} isHistory={isHistory} navigation={navigation} senderReview={senderReview} />
+        <VerifiedProof
+          record={record}
+          isHistory={isHistory}
+          senderReview={senderReview}
+          connectionLabel={connectionLabel}
+          onBackPressed={onBackPressed}
+          onGenerateNewPressed={onGenerateNewPressed}
+        />
       )}
-      {!(record.isVerified || senderReview) && <UnverifiedProof record={record} navigation={navigation} />}
+      {!(record.isVerified || senderReview) && (
+        <UnverifiedProof record={record} onBackPressed={onBackPressed} onGenerateNewPressed={onGenerateNewPressed} />
+      )}
     </SafeAreaView>
   )
 }
